+++ gc-coffee/src/main/java/com/example/gccoffee/dto/product/ProductUpdateDto.java
@@ -4,5 +4,5 @@ import com.example.gccoffee.model.Category;
 
 import java.util.UUID;
 
-public record ProductUpdateDto (UUID productId, String productName, Category category, long price, String description) {
+public record ProductUpdateDto (UUID productId, String productName, Category category, int price, String description) {
 }
